+++ app/Config/Auth.php
@@ -36,11 +36,6 @@ class Auth extends \Myth\Auth\Config\Auth
 	 */
 	public $views = [
 		'login'		=> 'App\Views\layouts\login',
-		'register'		=> 'Myth\Auth\Views\register',
-		'forgot'		  => 'Myth\Auth\Views\forgot',
-		'reset'		   => 'Myth\Auth\Views\reset',
-		'emailForgot'	 => 'Myth\Auth\Views\emails\forgot',
-		'emailActivation' => 'Myth\Auth\Views\emails\activation',
 	];
 
 	/**
@@ -147,7 +142,8 @@ class Auth extends \Myth\Auth\Config\Auth
 	 *
 	 * @var string Name of the ActivatorInterface class
 	 */
-	public $requireActivation = 'Myth\Auth\Authentication\Activators\EmailActivator';
+	public $requireActivation = false;
+	//public $requireActivation = 'Myth\Auth\Authentication\Activators\EmailActivator';
 
 	/**
 	 * --------------------------------------------------------------------
@@ -174,7 +170,7 @@ class Auth extends \Myth\Auth\Config\Auth
 	 *
 	 * @var bool
 	 */
-	public $allowRemembering = false;
+	public $allowRemembering = true;
 
 	/**
 	 * --------------------------------------------------------------------