+++ app/Controllers/CDR/Dashboard.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace App\Controllers\CDR;
+use App\Controllers\BaseController;
+
+class Dashboard extends BaseController
+{
+	public function index()
+	{	
+        return view('layouts/template');
+	}
+}