+++ docs/changelog.md
@@ -1,3 +1,14 @@
+##### v0.0.8
+- Edited config auth in App\Config\Auth
+- Disabled auto routes in App\Config\Routes
+- Created routes for /
+- Removed unused View welcome_message
+- Created View header in Views\layouts
+- Created View sidebar in Views\layouts
+- Added html tags in Views\layouts\template
+##### v0.0.7
+- Created Seeder UserSeed
+- Update README.md
 ##### v0.0.6
 - Finished View layouts/login
 ##### v0.0.5